py_bot/requirements.txt
anyio==4.2.0
astroid==3.1.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
h11==0.14.0
html5lib==1.1
httpcore==1.0.3
httpx==0.26.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.0
pylint==3.1.0
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tomlkit==0.12.4
urllib3==2.2.0
webencodings==0.5.1
